requirements.txt
Faker==25.2.0
psycopg[binary]==3.1.18
python-dotenv==1.0.1